+++ src/main/kotlin/com/integrated/medical/records/domain/dto/EmergencyCardDTO.kt
@@ -6,7 +6,7 @@ import kotlin.streams.toList
 
 class EmergencyCardDTO(
 
-        val idEmergencyCardDTO: Long,
+        val idEmergencyCardDTO: Int,
 
         val declaration: String,
